Added plugin context view
- Added plugin context view
- Moved plugin type definition to separate file
- Added wip request forwarder
