+++ src/mod/plugins/plugins.go
@@ -1,48 +1,22 @@
 package plugins
 
+/*
+	Zoraxy Plugin Manager
+
+	This module is responsible for managing plugins
+	loading plugins from the disk
+	enable / disable plugins
+	and forwarding traffic to plugins
+*/
+
 import (
 	"errors"
-	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"sync"
 
-	_ "embed"
-
-	"imuslab.com/zoraxy/mod/database"
-	"imuslab.com/zoraxy/mod/dynamicproxy/dpcore"
-	"imuslab.com/zoraxy/mod/info/logger"
-	zoraxyPlugin "imuslab.com/zoraxy/mod/plugins/zoraxy_plugin"
 	"imuslab.com/zoraxy/mod/utils"
 )
-
-type Plugin struct {
-	RootDir string                   //The root directory of the plugin
-	Spec    *zoraxyPlugin.IntroSpect //The plugin specification
-	Enabled bool                     //Whether the plugin is enabled
-
-	//Runtime
-	AssignedPort int                  //The assigned port for the plugin
-	uiProxy      *dpcore.ReverseProxy //The reverse proxy for the plugin UI
-	process      *exec.Cmd            //The process of the plugin
-}
-
-type ManagerOptions struct {
-	PluginDir    string
-	SystemConst  *zoraxyPlugin.RuntimeConstantValue
-	Database     *database.Database
-	Logger       *logger.Logger
-	CSRFTokenGen func(*http.Request) string //The CSRF token generator function
-}
-
-type Manager struct {
-	LoadedPlugins sync.Map //Storing *Plugin
-	Options       *ManagerOptions
-}
-
-//go:embed no_img.png
-var noImg []byte
 
 // NewPluginManager creates a new plugin manager
 func NewPluginManager(options *ManagerOptions) *Manager {
