+++ src/web/components/gan.html
@@ -3,6 +3,10 @@
         <h2>Global Area Network</h2>
         <p>Virtual Network Hub that allows all networked devices to communicate as if they all reside in the same physical data center or cloud region</p>
     </div>
+    <div class="ui yellow message">
+        <b>Deprecation Notice</b>
+        <p>Global Area Network will be deprecating in v3.2.x and moved to Plugin</p>
+    </div>
     <div class="gansnetworks">
         <div class="ganstats ui basic segment">
             <div style="float: right; max-width: 300px; margin-top: 0.4em;">
